Migrated over the remaining commands. RH stayed.
